refactor(api): extract API-key scope guards + push org list filter to DB
Review follow-ups on #172:

1. Two inline `apiKey*ScopeGuard` middlewares duplicated between
   `routes/organizations.ts` and `routes/applications.ts` — extract both
   to `middleware/guards.ts` so the next app-scoped surface reuses the
   same pattern instead of growing a third variant.

2. `GET /api/orgs` filtered the org list in memory for API key callers,
   which meant the DB query still enumerated every org the key creator
   belongs to. Push the `orgId` predicate down to the query via a new
   optional `orgIdFilter` on `getUserOrganizations()` so a compromised
   key cannot cause cross-org enumeration reads.

No behavior change for sessions or for API keys — only the layer where
the filter runs. Existing test `GET /api/orgs returns only the key's
org` covers the observable contract.
